feat: conditionally render Add Wiki button based on hasPermission in WikiCard
