+++ [O
@@ -1,0 +1,6 @@
+Merge remote-tracking branch 'origin/main' into feature/wiki-card-with-links
+# Please enter a commit message to explain why this merge is necessary,
+# especially if it merges an updated upstream into a topic branch.
+#
+# Lines starting with '#' will be ignored, and an empty message aborts
+# the commit.
